Fixed all build warnings.

+++ SystemMonitor/Network/NetworkInfoController.m
@@ -778,13 +778,13 @@ static void reachabilityCallback(SCNetworkReachabilityRef target, SCNetworkReach
                     tp = (struct xtcpcb_n *)xgn;
                     break;
                 default:
-                    AMLogWarn(@"unknown kind %d", xgn->xgn_kind);
+                    AMLogWarn(@"unknown kind %ld", (long)xgn->xgn_kind);
                     break;
             }
         }
         else
         {
-            AMLogWarn(@"got %d twice.", xgn->xgn_kind);
+            AMLogWarn(@"got %ld twice.", (long)xgn->xgn_kind);
         }
         
         if ((connectionType == CONNECTION_TYPE_TCP4 && which != ALL_XGN_KIND_TCP) ||
